Renamed color_utils to isValidColor
