+++ packages/components/src/Form/Inputs/InputColor/utils/isValidColor.ts
